Renamespace MachineRequest
