+++ src/Model/MachineRequestInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Model\ApiRequest;
+namespace App\Model;
 
 interface MachineRequestInterface
 {
