+++ src/Services/MachineHandler/CreateMachineHandler.php
@@ -7,10 +7,10 @@ namespace App\Services\MachineHandler;
 use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
 use App\MessageDispatcher\MachineRequestMessageDispatcherInterface;
-use App\Model\ApiRequest\MachineRequest;
-use App\Model\ApiRequest\MachineRequestInterface;
 use App\Model\ApiRequestOutcome;
 use App\Model\MachineProviderActionInterface;
+use App\Model\MachineRequest;
+use App\Model\MachineRequestInterface;
 use App\Model\Worker\State;
 use App\Repository\MachineRepository;
 use App\Services\ApiActionRetryDecider;
