+++ src/Services/MachineHandler/UpdateWorkerHandler.php
@@ -7,9 +7,9 @@ namespace App\Services\MachineHandler;
 use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
 use App\MessageDispatcher\MachineRequestMessageDispatcherInterface;
-use App\Model\ApiRequest\MachineRequestInterface;
 use App\Model\ApiRequestOutcome;
 use App\Model\MachineProviderActionInterface;
+use App\Model\MachineRequestInterface;
 use App\Model\Worker\State;
 use App\Model\Worker\StateTransitionSequence;
 use App\Repository\MachineRepository;
