+++ tests/Functional/Services/MachineHandler/CreateMachineHandlerTest.php
@@ -7,8 +7,8 @@ namespace App\Tests\Functional\Services\MachineHandler;
 use App\Exception\MachineProvider\Exception;
 use App\Exception\UnsupportedProviderException;
 use App\Message\MachineRequestMessage;
-use App\Model\ApiRequest\MachineRequest;
 use App\Model\MachineProviderActionInterface;
+use App\Model\MachineRequest;
 use App\Model\ProviderInterface;
 use App\Model\Worker\State;
 use App\Services\ExceptionLogger;
