+++ tests/Mock/MessageHandler/MockUpdateWorkerHandler.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Mock\MessageHandler;
 
-use App\Model\ApiRequest\MachineRequestInterface;
+use App\Model\MachineRequestInterface;
 use App\Services\MachineHandler\UpdateWorkerHandler;
 use Mockery\MockInterface;
