+++ tests/Unit/Model/CreateMachineRequestTest.php
@@ -12,7 +12,7 @@ class CreateMachineRequestTest extends TestCase
     {
         $workerId = '123';
 
-        $request = new \App\Model\ApiRequest\MachineRequest($workerId);
+        $request = new \App\Model\MachineRequest($workerId);
         self::assertSame($workerId, $request->getWorkerId());
         self::assertSame(0, $request->getRetryCount());
     }
